fix: rename data provider
