+++ src/main/kotlin/it/pagopa/ecommerce/helpdesk/dataproviders/v1/mongo/PmTransactionHistoryDataProvider.kt
@@ -12,7 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired
 import reactor.core.publisher.Flux
 import reactor.core.publisher.Mono
 
-class PmTransactionDataProvider(
+class PmTransactionHistoryDataProvider(
     @Autowired private val pmTransactionsRepository: PmTransactionsRepository
 ) : TransactionDataProvider {
